Reposition budgeting controls under category search
